Modify #omp parallel for parsing to work with PIR
1. Modifies how an OMPParallelForDirective object is formed. Mainly,
implicit variables that are needed by the OMP runtime are not generated anymore.
The final input to emitOMPParallelForDirective function is a regular C++
for loop.

2. This commit implements the initial version of parallel loops in PIR. It
creates fork/join instructions to fork the body of the loop as a separate
task.

3. In a parallel loop, instead of emitting a br to the increment block of the
loop (which is also the continuation task of the parallel region representing
the loop), a halt is emitted.

4. ParallelRegionInfo checks to see if a join instruction is dominated by its
corresponding fork instruction or not. This differs from Tapir where this
is not neccessary.